reorganize go code, add address index util
